+++ src/readthedocs_assistant/commands/__init__.py
@@ -1,0 +1,3 @@
+from .migrate_config import migrate_config
+
+__all__ = ["migrate_config"]
