+++ tests/test_github.py
@@ -3,10 +3,10 @@ from unittest import mock
 import pytest
 from gidgethub.abc import GitHubAPI
 
-from readthedocs_assistant.cli import fork_repo
+from readthedocs_assistant.github import fork_repo
 
 
-@mock.patch("readthedocs_assistant.cli.gidgethub.httpx.GitHubAPI", spec=GitHubAPI)
+@mock.patch("readthedocs_assistant.github.gidgethub.httpx.GitHubAPI", spec=GitHubAPI)
 @pytest.mark.asyncio
 async def test_fork_repo_awaits_fork_repo(mock_gh):
     target_repo = {"full_name": "owner/repo", "name": "repo"}
